# common-platform-org-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-org-mgmt
  description: Organization Management Core Application
  contact:
    name: Catalis Development
    email: dev@catalis.com
  version: 1.0.0
servers:
  - url: 'http://core.catalis.vc/organization'
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  '/api/v1/banks/{bankId}/calendars/{calendarId}/assignments/{assignmentId}':
    get:
      tags:
        - Calendar Assignment Management
      summary: Get calendar assignment by ID
      description: Returns an assignment of a specific working calendar based on its ID
      operationId: getCalendarAssignmentById
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: calendarId
          in: path
          description: ID of the calendar
          required: true
          schema:
            type: integer
            format: int64
        - name: assignmentId
          in: path
          description: ID of the assignment to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved calendar assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarAssignmentDTO'
        '404':
          description: 'Bank, calendar, or assignment not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarAssignmentDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarAssignmentDTO'
    put:
      tags:
        - Calendar Assignment Management
      summary: Update calendar assignment
      description: >-
        Updates an existing assignment of a specific working calendar with the
        provided details
      operationId: updateCalendarAssignment
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: calendarId
          in: path
          description: ID of the calendar
          required: true
          schema:
            type: integer
            format: int64
        - name: assignmentId
          in: path
          description: ID of the assignment to update
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarAssignmentDTO'
        required: true
      responses:
        '200':
          description: Calendar assignment successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarAssignmentDTO'
        '400':
          description: Invalid calendar assignment data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarAssignmentDTO'
        '404':
          description: 'Bank, calendar, or assignment not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarAssignmentDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarAssignmentDTO'
    delete:
      tags:
        - Calendar Assignment Management
      summary: Delete calendar assignment
      description: Deletes an assignment of a specific working calendar based on its ID
      operationId: deleteCalendarAssignment
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: calendarId
          in: path
          description: ID of the calendar
          required: true
          schema:
            type: integer
            format: int64
        - name: assignmentId
          in: path
          description: ID of the assignment to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Calendar assignment successfully deleted
        '404':
          description: 'Bank, calendar, or assignment not found'
        '500':
          description: Internal server error
  '/api/v1/banks/{bankId}/calendars/{calendarId}':
    get:
      tags:
        - Working Calendar Management
      summary: Get working calendar by ID
      description: Returns a working calendar of a specific bank based on its ID
      operationId: getWorkingCalendarById
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: calendarId
          in: path
          description: ID of the calendar to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved working calendar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkingCalendarDTO'
        '404':
          description: Bank or calendar not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkingCalendarDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkingCalendarDTO'
    put:
      tags:
        - Working Calendar Management
      summary: Update working calendar
      description: >-
        Updates an existing working calendar of a specific bank with the
        provided details
      operationId: updateWorkingCalendar
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: calendarId
          in: path
          description: ID of the calendar to update
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkingCalendarDTO'
        required: true
      responses:
        '200':
          description: Working calendar successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkingCalendarDTO'
        '400':
          description: Invalid working calendar data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkingCalendarDTO'
        '404':
          description: Bank or calendar not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkingCalendarDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkingCalendarDTO'
    delete:
      tags:
        - Working Calendar Management
      summary: Delete working calendar
      description: Deletes a working calendar of a specific bank based on its ID
      operationId: deleteWorkingCalendar
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: calendarId
          in: path
          description: ID of the calendar to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Working calendar successfully deleted
        '404':
          description: Bank or calendar not found
        '500':
          description: Internal server error
...
  '/api/v1/banks/{bankId}/calendars/{calendarId}/assignments/filter':
    post:
      tags:
        - Calendar Assignment Management
      summary: Get all assignments for a calendar with filtering
      description: >-
        Returns a paginated list of assignments for a specific working calendar
        based on filter criteria
      operationId: filterCalendarAssignments
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: calendarId
          in: path
          description: ID of the calendar
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestCalendarAssignmentDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved calendar assignments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter criteria supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseCalendarAssignmentDTO'
        '404':
          description: Bank or calendar not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseCalendarAssignmentDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseCalendarAssignmentDTO'
  '/api/v1/banks/{bankId}/calendars/{calendarId}/assignments':
    post:
      tags:
        - Calendar Assignment Management
      summary: Create a new assignment for a calendar
      description: >-
        Creates a new assignment for a specific working calendar with the
        provided details
      operationId: createCalendarAssignment
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: calendarId
          in: path
          description: ID of the calendar
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarAssignmentDTO'
        required: true
      responses:
        '201':
          description: Calendar assignment successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarAssignmentDTO'
        '400':
          description: Invalid calendar assignment data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarAssignmentDTO'
        '404':
          description: Bank or calendar not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarAssignmentDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarAssignmentDTO'
  '/api/v1/banks/{bankId}/calendars/filter':
    post:
      tags:
        - Working Calendar Management
      summary: Get all calendars for a bank with filtering
      description: >-
        Returns a paginated list of working calendars for a specific bank based
        on filter criteria
      operationId: filterWorkingCalendars
      parameters:
        - name: bankId
          in: path
          description: ID of the bank
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestWorkingCalendarDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved working calendars
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter criteria supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseWorkingCalendarDTO'
        '404':
          description: Bank not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseWorkingCalendarDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseWorkingCalendarDTO'
...
components:
  schemas:
...
    CalendarAssignmentDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        calendarId:
          type: integer
          format: int64
        branchId:
          type: integer
          format: int64
        departmentId:
          type: integer
          format: int64
        positionId:
          type: integer
          format: int64
        effectiveFrom:
          type: string
          format: date-time
        effectiveTo:
          type: string
          format: date-time
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: integer
          format: int64
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: integer
          format: int64
    WorkingCalendarDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        bankId:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        isDefault:
          type: boolean
        timeZoneId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: integer
          format: int64
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: integer
          format: int64
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: 'The direction of sorting, either ASC or DESC.'
          example: DESC
          default: DESC
      description: >-
        Represents a pagination request for retrieving paginated results,
        including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.
...
    FilterRequestCalendarAssignmentDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/CalendarAssignmentDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for calendar assignments
    PaginationResponseCalendarAssignmentDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/CalendarAssignmentDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.
    FilterRequestWorkingCalendarDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/WorkingCalendarDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for working calendars
    PaginationResponseWorkingCalendarDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/WorkingCalendarDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.
...